hackproject/src/api/requirements.txt
# ============================================================================
# Python Requirements para Azure Functions
# ============================================================================
# BiciMAD Low Emission Router Backend APIs
# ============================================================================

# Azure Functions Core
azure-functions>=1.18.0
azure-identity>=1.15.0

# Azure SDKs
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# HTTP Clients
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
geopy>=2.4.1

# JSON & Data Validation
pydantic>=2.5.3
python-dotenv>=1.0.0

# Caching
redis>=5.0.1

# Date/Time
python-dateutil>=2.8.2

# Testing (Dev only)
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx[http2]>=0.26.0

# Code Quality (Dev only)
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0

# Application Insights
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1
